third_party/WebKit/LayoutTests/http/tests/devtools/console/console-copy-treeoutline-expected.txt

Tests that console copies tree outline messages properly.


Running: testSelectAll
Selected text: console-copy-treeoutline.js@evaluation0:16 {foo: 1, bar: "string"}
